+++ server/src/gps.ts
@@ -1,8 +1,7 @@
-import { requestJson } from 'by-request';
 import { ChildProcess } from 'child_process';
 import { parseISODate } from '@tubular/time';
 import { abs, floor, max, round } from '@tubular/math';
-import { asLines, processMillis } from '@tubular/util';
+import { asLines, processMillis, toNumber } from '@tubular/util';
 import { NtpData } from './ntp-data';
 import { ErrorMode, monitorProcess, spawn } from './process-util';
 import { ForecastData, GpsData, TimeInfo } from './shared-types';
@@ -10,6 +9,7 @@ import { TaiUtc } from './tai-utc';
 import { TimePoller } from './time-poller';
 import { roughDistanceBetweenLocationsInKm, timeStamp, unref } from './awcs-util';
 import { getForecast } from './weatherbit-forecast';
+import { requestJson } from './request-cache';
 
 const BILLION = BigInt('1000000000');
 const THOUSAND = BigInt('1000');
@@ -18,6 +18,7 @@ const CLOCK_CHECK = 30_000; // Half minute
 const GPS_DROPOUT_MAX = 60_0000; // 1 minute
 const CHECK_LOCATION_RETRY_DELAY = 300_000; // 5 minutes
 const OVER_QUOTA_RETRY_DELAY = 86_400_000; // 1 day
+const LOCATION_CACHE_TIME = 86_400; // 1 day (in secs)
 
 export class Gps extends TimePoller {
   private clockCheckTimeout: any;
@@ -61,6 +62,7 @@ export class Gps extends TimePoller {
 
   public close(): void {
     this.gpspipe.kill('SIGINT');
+    this.lastGpsInfo = null;
 
     if (this.clockCheckTimeout)
       clearTimeout(this.clockCheckTimeout);
@@ -97,8 +99,8 @@ export class Gps extends TimePoller {
         const obj = JSON.parse(data.toString());
 
         if (obj?.class === 'TPV' && obj.lat != null) {
-          this.gpsData.latitude = obj.lat;
-          this.gpsData.longitude = obj.lon;
+          this.gpsData.latitude = toNumber(obj.lat.toFixed(5));
+          this.gpsData.longitude = toNumber(obj.lon.toFixed(5));
           this.gpsData.altitude = obj.altHAE ?? obj.alt ?? 0;
           this.gpsData.fix = obj.status ?? this.gpsData.fix ?? 0;
           this.gpsData.pps = this.systemTimeIsGps;
@@ -204,7 +206,7 @@ export class Gps extends TimePoller {
       this.gpsData.signalQuality = 0;
     }
 
-    if (this.lastGpsInfo < 0)
+    if (this.lastGpsInfo != null && this.lastGpsInfo < 0)
       this.monitorGps();
 
     this.clockCheckTimeout = unref(setTimeout(() => {
@@ -251,7 +253,7 @@ export class Gps extends TimePoller {
     try {
       const url = `https://maps.googleapis.com/maps/api/geocode/json?key=${this.googleKey}` +
         `&result_type=locality|administrative_area_level_3&latlng=${coords.latitude},${coords.longitude}`;
-      const data = await requestJson(url);
+      const data = await requestJson(LOCATION_CACHE_TIME, url);
 
       if (data?.status === 'OK' && data.results?.length > 0)
         coords.city = data.results[0].formatted_address;
@@ -274,7 +276,7 @@ export class Gps extends TimePoller {
     try {
       const url = `https://maps.googleapis.com/maps/api/timezone/json?key=${this.googleKey}` +
         `&location=${coords.latitude},${coords.longitude}&timestamp=${floor(Date.now() / 1000)}`;
-      const data = await requestJson(url);
+      const data = await requestJson(LOCATION_CACHE_TIME, url);
 
       if (data?.status === 'OK' && data.timeZoneId)
         coords.timezone = data.timeZoneId;
@@ -283,7 +285,7 @@ export class Gps extends TimePoller {
 
         if (data.status === 'REQUEST_DENIED')
           this.googleAccessDenied = true;
-        else if (data.status === 'OVER_DAILY_LIMIT')
+        else if (data.status === 'OVER_DAILY_LIMIT' || data.errorMessage.includes(' exceeded your daily request quota '))
           this.checkLocationRetry = now + OVER_QUOTA_RETRY_DELAY;
         else
           this.checkLocationRetry = now + CHECK_LOCATION_RETRY_DELAY;
@@ -292,6 +294,9 @@ export class Gps extends TimePoller {
     catch (err) {
       this.checkLocationRetry = now + CHECK_LOCATION_RETRY_DELAY;
       console.error('%s -- Google timezone check:', timeStamp(), err);
+
+      if ((err.message ?? err).toString().includes(' exceeded your daily request quota '))
+        this.checkLocationRetry = now + OVER_QUOTA_RETRY_DELAY;
     }
   }
 
@@ -304,7 +309,7 @@ export class Gps extends TimePoller {
       if (this.weatherbitKey)
         forecast = await getForecast({ query: { lat, lon, co: 'true' } } as any);
       else
-        forecast = await requestJson(`https://weather.shetline.com/wbproxy?lat=${lat}&lon=${lon}&co=true`);
+        forecast = await requestJson(LOCATION_CACHE_TIME, `https://weather.shetline.com/wbproxy?lat=${lat}&lon=${lon}&co=true`);
     }
     catch {
       this.checkLocationRetry = now + CHECK_LOCATION_RETRY_DELAY;